Bandshift, calibration and polarization angle systematics added to maps2params (#99)
* Codes modified to add band integration and band shift, calibration and polarization angle systematics. Parameter file and the possibility to set the seed for the simulations
 added.

* Corrections to the parameter files and codes

* Modified strings and restored compatibility with the case exp != "LAT" in get_maps_to_params_spectra.py

+++ project/maps2params/python/maps_to_params_multifrequency_covmat.py
@@ -88,8 +88,8 @@ for sid, name in enumerate(spec_name):
         if (na == nb) & (spec == 'ET'):
             block_to_delete = np.append(block_to_delete, np.arange(id_start, id_stop))
 
-full_analytic_cov = np.delete(full_analytic_cov, block_to_delete, axis=1)
-full_analytic_cov = np.delete(full_analytic_cov, block_to_delete, axis=0)
+full_analytic_cov = np.delete(full_analytic_cov, block_to_delete.astype(int), axis=1)
+full_analytic_cov = np.delete(full_analytic_cov, block_to_delete.astype(int), axis=0)
 
 np.save("%s/truncated_analytic_cov.npy"%cov_dir, full_analytic_cov)
